requirements.txt
openai-agents
python-pptx
fastapi
uvicorn
httpx
